Fix memory corruption and cleanup warnings.

+++ src/expr/eeval.cc
@@ -190,9 +190,9 @@ evaluate_expr::operator()(const binary_op_expr *p) const
 		case binary_op_expr::LOGIC_NXOR:
 			logic_xnor(*p->_result, left, right); break;
 		case binary_op_expr::LEFT_SHIFT:
-			binary_lshf(*p->_result, left, static_cast<unsigned long>(right)); break;
+			binary_lshf(*p->_result, left, right); break;
 		case binary_op_expr::RIGHT_SHIFT:
-			binary_rshf(*p->_result, left, static_cast<unsigned long>(right)); break;
+			binary_rshf(*p->_result, left, right); break;
 		}
 	return *p->_result;
 	}
